Add -include flag
This patch adds an -include flag which takes a space-separated string of
package paths that should be considered against the globs provided to
`-copy`.

The use-case encountered in my own context was converting a project that
used `github.com/tensorflow/tensorflow`, which `go mod` fails to record
the `github.com/tensorflow/tensorflow/tensorflow/c` folder that contains
the header files necessary for compilation.