add flag that copies additional directories into ./vendor
In the case of a package like `github.com/prometheus/client_model`,
there are protobuf files in `github.com/prometheus/client_model` which
do not get copied at all even with this tool as it is not specified in
`./vendor/modules.txt`.

Add the ability to do this by adding an `-include` flag, which
takes in `<dir_to_copy>` fields separated by commas.

(Example of this in CockroachDB: cockroachdb/cockroach#49447)